Should hopefully fix the errors.
For some reason, other people could not launch the tool. However, Tom could launch but not get past the CheckFirstRun which was weird. I have changed the ServicePoint locations and improved the ping and locking statements. Please let this work, the most annoying thing is that it works fine for me but not for others like HUH.

The change to the NewVersion attribute and NLinkAnon is the changes with this commit, the changes mentioned above are contained within the new update.